fix(generator): revert to raw xxh3 IDs for native parity
The generateWebSafeId wrapper in the generator produced IDs that
didn't match the native Rust core's raw xxh3_64 computation, causing
1181 test failures ("Collection id is invalid").

- Revert ObjectInfo.id, ObjectIndex.id, ObjectLink.id to raw xxh3
- Move generateWebSafeId to web runtime for future web support
- Move web_safe_id_test to isar_community package
